refactor: currencies and currency code move to utils file from editor component, feat: currency file on utils
